+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Apply the new large headline design to Agreements. ERM-260.
 * Display coverage for Agreement lines on Create Agreement pane. ERM-168.
 * Added ability to manage amendments for linked licenses. ERM-428 ERM-429
+* Added agreement line active from/to dates. ERM-456
 
 ## 3.2.4 2019-09-10
 * Add support for interface `invoice` version `2.0`.